Catch all zero diff_methods in Hessian
Fixes a bug in the param_shift_hessian where the tape execution results
are empty, due to an optimization which prevents all device execution when
all elements of the Hessian are known to be 0.
The post-processing functions of the parameter-shift transforms however
assume that the tape results contain atleast one element.